requirements.txt
torch>=1.9.0
transformers>=4.11.0
pydantic>=1.8.2
networkx>=2.6.3
neo4j>=4.4.0
pyyaml>=5.4.1
tqdm>=4.62.3
simpletransformers  
spacy 
spacy-transformers
